Refactor: Use long for damage to support higher values
This change modifies various parts of the code to use the `long` data type for damage calculations and storage. This allows for significantly higher damage values, preventing overflow issues and enabling more powerful attacks and mob stats. The changes include:

- **`BotAttackplayer.java`**: Updated `realDame` to `long` and used `writeLong` for sending damage data.
- **`BanDoKhoBau.java`, `DestronGas.java`, `RedRibbonHQ.java`, `SuperDivineWaterService.java`**: Removed damage and HP caps for mobs, allowing for higher values.
- **`Mob.java`**: Changed `damage` to `long` in `injured` method and updated `sendMobDieAffterAttacked` and `sendMobStillAliveAffterAttacked` to use `writeLong`. Also changed `dameMob` to `long` in `attack` method and `sendMobAttackMe` to use `writeLong`.
- **`MobMe.java`**: Updated `dame` to `long` and used `writeLong` for sending damage data.
- **`MobPoint.java`**: Changed `dame` to `long` and updated `getDameAttack` to return `long`.
- **`GaChinCua.java`, `GauTuongCuop.java`, `Hirudegarn.java`, `NguaChinLmao.java`, `Piano.java`, `RobotBaoVe.java`, `VoiChinNga.java`, `VuaBachTuoc.java`**: Updated damage calculations and sending to use `long`.
- **`NPoint.java`**: Changed `dameAfter`, `dame`, `dameg`, and `dameAdd` to `long`. Removed damage cap and updated `getDameAttack` to return `long`.
- **`Player.java`**: Changed `lastDamage` to `long` and updated `injured` method to return `long`.
- **`Service.java`**: Updated `sendMessAllPlayerInMap` to use `writeLong` for damage.
- **`SkillService.java`**: Updated damage calculations and sending to use `long`.

+++ Source By Mr Blue/src/nro/models/Bot/BotAttackplayer.java
@@ -152,7 +152,7 @@ public class BotAttackplayer extends Bot {
         }
 
         boolean isMiss = false;
-        int realDame = targetPlayer.injured(this, isMiss ? 0 : damage, false, false);
+        long realDame = targetPlayer.injured(this, isMiss ? 0 : damage, false, false);
 
         Skill skill = playerSkill.skillSelect;
 
@@ -164,7 +164,7 @@ public class BotAttackplayer extends Bot {
             msg.writer().writeInt((int) targetPlayer.id);
             msg.writer().writeByte(1);
             msg.writer().writeByte(0);
-            msg.writer().writeInt(realDame);
+            msg.writer().writeLong(realDame);  // Dùng long cho damage cao
             msg.writer().writeBoolean(targetPlayer.isDie());
             msg.writer().writeBoolean(nPoint.isCrit);
             Service.gI().sendMessAllPlayerInMap(this, msg);
