+++ Source By Mr Blue/src/nro/models/mob_bigboss/RobotBaoVe.java
@@ -24,8 +24,8 @@ public class RobotBaoVe extends BigBoss {
         Player player = getPlayerCanAttack();
         if (!isDie() && !effectSkill.isHaveEffectSkill() && Util.canDoWithTime(lastBigBossAttackTime, 1000)) {
             if (player != null) {
-                int dameMob = this.point.getDameAttack();
-                int dame = player.injured(null, dameMob, false, true);
+                long dameMob = this.point.getDameAttack();
+                long dame = player.injured(null, dameMob, false, true);
                 action = 0;
                 int dis = Util.getDistance(player, this);
                 if (dis <= 100) {
